move api_workflow_artifacts into new file
