+++ lightly/api/__init__.py
@@ -4,7 +4,7 @@
 # All Rights Reserved
 
 from lightly.api.api_workflow_client import ApiWorkflowClient
-from lightly.api.api_workflow_compute_worker import ArtifactNotExist
+from lightly.api.api_workflow_artifacts import ArtifactNotExist
 from lightly.api.patch_rest_client import patch_rest_client
 
 from lightly.openapi_generated.swagger_client.rest import RESTClientObject
