+++ lightly/api/api_workflow_client.py
@@ -6,6 +6,7 @@ import os
 
 import requests
 
+from lightly.api.api_workflow_artifacts import _AritfactsMixin
 from lightly.api.api_workflow_predictions import _PredictionsMixin
 from lightly.api.api_workflow_tags import _TagsMixin
 from requests import Response
@@ -59,6 +60,7 @@ class ApiWorkflowClient(_UploadEmbeddingsMixin,
                         _ComputeWorkerMixin,
                         _CollaborationMixin,
                         _PredictionsMixin,
+                        _AritfactsMixin,
                         ):
     """Provides a uniform interface to communicate with the Lightly API.
